book

+++ book-service/src/main/java/com/example/bookservice/service/BookService.java
@@ -1,8 +1,8 @@
 package com.example.bookservice.service;
 
+import com.example.bookservice.kafka.BookProducer;
 import com.example.bookservice.model.Book;
 import com.example.bookservice.repository.BookRepository;
-import com.logging.LoggerService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,10 +12,11 @@ import java.util.Optional;
 public class BookService {
 
     private final BookRepository bookRepository;
-    private final LoggerService logger = LoggerService.getInstance();
+    private final BookProducer bookProducer;
 
-    public BookService(BookRepository bookRepository) {
+    public BookService(BookRepository bookRepository, BookProducer bookProducer) {
         this.bookRepository = bookRepository;
+        this.bookProducer = bookProducer;
     }
 
     public List<Book> getAllBooks() {
@@ -26,12 +27,38 @@ public class BookService {
         return bookRepository.findById(id);
     }
 
-    public Book saveBook(Book book) {
-        logger.info("Created new book " + book.toString());
-        return bookRepository.save(book);
+    public Book createBook(Book book) {
+        book.setAvailable(true); // By default, the book is available
+        Book savedBook = bookRepository.save(book);
+        bookProducer.sendBookEvent("Book created: " + savedBook.getTitle());
+        return savedBook;
     }
 
-    public void deleteBook(Long id) {
-        bookRepository.deleteById(id);
+    public boolean deleteBook(Long id) {
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+            bookProducer.sendBookEvent("Book deleted with ID: " + id);
+            return true;
+        }
+        return false;
+    }
+
+    public Optional<Book> updateAvailability(Long id, boolean available) {
+        return bookRepository.findById(id).map(book -> {
+            book.setAvailable(available);
+            Book updatedBook = bookRepository.save(book);
+            return updatedBook;
+        });
+    }
+
+    public Optional<Book> updateBook(Long id, Book updatedBook) {
+        return bookRepository.findById(id).map(existingBook -> {
+            existingBook.setTitle(updatedBook.getTitle());
+            existingBook.setAuthor(updatedBook.getAuthor());
+            existingBook.setAvailable(updatedBook.isAvailable());
+            Book savedBook = bookRepository.save(existingBook);
+            bookProducer.sendBookEvent("Book updated: " + savedBook.getTitle());
+            return savedBook;
+        });
     }
 }
